docs(economy): add the relevant documentation to functions to clear up their purpose
